Extracted trigger logic from TextArea to TriggerTextArea
